Reduce sorting code duplication
Fix missing TpLoginRequiredMixin
