+++ trustpoint/pki/views/domains.py
@@ -16,7 +16,13 @@ from django.views.generic.edit import FormView
 from pki.forms import DevIdRegistrationForm, DevIdAddMethodSelectForm
 from pki.models import DomainModel, DevIdRegistration
 from pki.models.truststore import TruststoreModel
-from trustpoint.views.base import ContextDataMixin, TpLoginRequiredMixin, BulkDeleteView
+from trustpoint.views.base import (
+    ContextDataMixin,
+    TpLoginRequiredMixin,
+    BulkDeleteView,
+    ListInDetailView,
+    SortableTableMixin
+)
 
 
 class PkiProtocol(enum.Enum):
@@ -35,7 +41,7 @@ class DomainContextMixin(ContextDataMixin):
     context_page_name = 'domains'
 
 
-class DomainTableView(DomainContextMixin, ListView):
+class DomainTableView(DomainContextMixin, TpLoginRequiredMixin, SortableTableMixin, ListView):
     """Domain Table View."""
 
     model = DomainModel
@@ -43,27 +49,6 @@ class DomainTableView(DomainContextMixin, ListView):
     context_object_name = 'domain-new'
     paginate_by = 5  # Number of items per page
     default_sort_param = 'unique_name'
-
-    def get_queryset(self):
-        queryset = self.model.objects.all()
-
-        # Get sort parameter (e.g., "name" or "-name")
-        sort_param = self.request.GET.get('sort', self.default_sort_param)
-        return queryset.order_by(sort_param)
-
-    def get_context_data(self, **kwargs) -> dict:
-        context = super().get_context_data(**kwargs)
-
-        # Get current sorting column
-        sort_param = self.request.GET.get('sort', self.default_sort_param)
-        is_desc = sort_param.startswith('-')  # Check if sorting is descending
-
-        # Pass sorting details to the template
-        context.update({
-            'current_sort': sort_param,
-            'is_desc': is_desc,
-        })
-        return context
 
 
 class DomainCreateView(DomainContextMixin, TpLoginRequiredMixin, CreateView):
@@ -84,62 +69,12 @@ class DomainUpdateView(DomainContextMixin, TpLoginRequiredMixin, UpdateView):
     ignore_url = reverse_lazy('pki:domains')
 
 
-class ListInDetailView(ListView):
-    """Helper view that combines a DetailView and a ListView.
-
-    This is useful for displaying a list within a DetailView.
-    Note that 'model' and 'context_object_name' refer to the ListView.
-    Use 'detail_model' and 'detail_context_object_name' for the DetailView.
-    """
-    detail_context_object_name = 'object'
-
-    def get(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        return super().get(request, *args, **kwargs)
-
-    def get_queryset_for_object(self):
-        return self.detail_model.objects.all()
-
-    def get_object(self):
-        queryset = self.get_queryset_for_object()
-        pk = self.kwargs.get('pk')
-        if pk is None:
-            raise AttributeError('pk expected in url')
-        return get_object_or_404(queryset, pk=pk)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context[self.detail_context_object_name] = self.object
-        return context
-
-
-class DomainDevIdRegistrationTableMixin(ListInDetailView):
+class DomainDevIdRegistrationTableMixin(SortableTableMixin, ListInDetailView):
 
     model = DevIdRegistration
     paginate_by = 5  # Number of items per page
     context_object_name = 'devid_registrations'
     default_sort_param = 'unique_name'
-    
-    def get_queryset(self):
-        queryset = self.model.objects.filter(domain=self.get_object())
-
-        # Get sort parameter (e.g., "name" or "-name")
-        sort_param = self.request.GET.get('sort', self.default_sort_param)
-        return queryset.order_by(sort_param)
-
-    def get_context_data(self, **kwargs) -> dict:
-        context = super().get_context_data(**kwargs)
-
-        # Get current sorting column
-        sort_param = self.request.GET.get('sort', self.default_sort_param)
-        is_desc = sort_param.startswith('-')  # Check if sorting is descending
-
-        # Pass sorting details to the template
-        context.update({
-            'current_sort': sort_param,
-            'is_desc': is_desc,
-        })
-        return context
 
 
 class DomainConfigView(DomainContextMixin, TpLoginRequiredMixin, DomainDevIdRegistrationTableMixin, ListInDetailView):
@@ -178,11 +113,11 @@ class DomainConfigView(DomainContextMixin, TpLoginRequiredMixin, DomainDevIdRegi
         return HttpResponseRedirect(self.success_url)
 
 
-class DomainDetailView(DomainContextMixin, TpLoginRequiredMixin, DomainDevIdRegistrationTableMixin, DetailView):
+class DomainDetailView(DomainContextMixin, TpLoginRequiredMixin, DomainDevIdRegistrationTableMixin, ListInDetailView):
 
-    model = DomainModel
+    detail_model = DomainModel
     template_name = 'pki/domains/details.html'
-    context_object_name = 'domain'
+    detail_context_object_name = 'domain'
 
 
 class DomainCaBulkDeleteConfirmView(DomainContextMixin, TpLoginRequiredMixin, BulkDeleteView):
