+++ trustpoint/trustpoint/views/base.py
@@ -17,10 +17,11 @@ from django.http import Http404
 from django.core.exceptions import ImproperlyConfigured
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.http import HttpRequest, HttpResponseRedirect
+from django.shortcuts import get_object_or_404
 from django.utils.translation import gettext_lazy as _
 from django.views.generic.base import RedirectView
 from django.views.generic.edit import FormMixin
-from django.views.generic.list import BaseListView, MultipleObjectTemplateResponseMixin
+from django.views.generic.list import BaseListView, ListView, MultipleObjectTemplateResponseMixin
 
 from typing import TYPE_CHECKING
 
@@ -32,6 +33,63 @@ class IndexView(RedirectView):
 
     permanent: bool = False
     pattern_name: str = 'pki:certificates'
+
+
+class ListInDetailView(ListView):
+    """Helper view that combines a DetailView and a ListView.
+
+    This is useful for displaying a list within a DetailView.
+    Note that 'model' and 'context_object_name' refer to the ListView.
+    Use 'detail_model' and 'detail_context_object_name' for the DetailView.
+    """
+    detail_context_object_name = 'object'
+
+    def get(self, request, *args, **kwargs):
+        self.object = self.get_object()
+        return super().get(request, *args, **kwargs)
+
+    def get_queryset_for_object(self):
+        return self.detail_model.objects.all()
+
+    def get_object(self):
+        queryset = self.get_queryset_for_object()
+        pk = self.kwargs.get('pk')
+        if pk is None:
+            exc_msg = 'detail object pk expected in url'
+            raise AttributeError(exc_msg)
+        return get_object_or_404(queryset, pk=pk)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context[self.detail_context_object_name] = self.object
+        return context
+    
+
+class SortableTableMixin:
+    """Adds utility for sorting a ListView query by URL parameters
+
+    default_sort_param must be set in the view to specify default sorting order.
+    """
+    def get_queryset(self):
+        queryset = self.model.objects.all()
+
+        # Get sort parameter (e.g., "name" or "-name")
+        sort_param = self.request.GET.get('sort', self.default_sort_param)
+        return queryset.order_by(sort_param)
+
+    def get_context_data(self, **kwargs) -> dict:
+        context = super().get_context_data(**kwargs)
+
+        # Get current sorting column
+        sort_param = self.request.GET.get('sort', self.default_sort_param)
+        is_desc = sort_param.startswith('-')  # Check if sorting is descending
+
+        # Pass sorting details to the template
+        context.update({
+            'current_sort': sort_param,
+            'is_desc': is_desc,
+        })
+        return context
 
 
 class TpLoginRequiredMixin(LoginRequiredMixin):
